Cride/requirements/base.txt
# Base
pytz==2018.7
python-slugify==1.2.6
Pillow==5.3.0
psycopg2==2.7.4 --no-binary psycopg2

# Django
django==3.0.0

# Environment
django-environ==0.4.5

# Passwords security
argon2-cffi==20.1.0

# Static files
whitenoise==4.1.2

# Django Model Utils
django-model-utils